+++ packages/uni_translate/lib/uni_translate.dart
@@ -5,6 +5,7 @@ export 'package:translation_engine_caiyun/translation_engine_caiyun.dart';
 export 'package:translation_engine_deepl/translation_engine_deepl.dart';
 export 'package:translation_engine_google/translation_engine_google.dart';
 export 'package:translation_engine_iciba/translation_engine_iciba.dart';
+export 'package:translation_engine_openai/translation_engine_openai.dart';
 export 'package:translation_engine_tencent/translation_engine_tencent.dart';
 export 'package:translation_engine_youdao/translation_engine_youdao.dart';
 export 'package:uni_translate_client/uni_translate_client.dart';