+++ packages/uni_translate/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.3
+
+* bump uni_translate_client to 0.1.3
+* add openai translation engine
+
 ## 0.1.2
 
 * chore: bump dart to 3.0.0 & fix lint issues