+++ packages/uni_translate_client/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.3
+
+* feat: Support for streaming output
+
 ## 0.1.2
 
 * chore: bump dart to 3.0.0 & fix lint issues